% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTypeR.R
\name{dataTypeR}
\alias{dataTypeR}
\title{Profile Column Data Types}
\usage{
dataTypeR(df)
}
\arguments{
\item{data}{An R object to be coerced into a data frame}
}
\value{
data.frame Columns tested for various data types and the percentage
likelihood of each data type tested
}
\description{
Tests each column in the data frame against
a number of regular expression conditions in
order to suggest the \strong{most likely}
overall data type
}
\details{
The regular expression tests employed search for possible
dates, times, integers, doubles and null values.
All of the various datetime formats are tested and if one
is found to match then the exact format code is returned.
If none of the above matches are found then the column is
assumed to be a character column.
}
\examples{

library(nycflights13)
dataTypeR(flights)
}
